requirements.txt
pyserial
mysql-connector